completed exercise 1 Css Methods
Created new file style.css in the current working dir of ex 1
inside which i have added the syte for div element.
Modified the index.html file to link the external css file.
Added the style tag for the P element with internal method.
Added the inline CSS for the button element.